+++ src/pages/order-confirm/order-contents/pay-contents/DeliveryInfo.tsx
@@ -1,12 +1,12 @@
 import styled from "@emotion/styled";
 import { useState } from "react";
-import { getShoppingCartData } from "../../../api/cart";
-import { Flex } from "../../../components/common";
-import CheckboxLabel from "../../../components/common/CheckboxLabel";
-import { useAPIDataContext } from "../../../context/APIDataProvider";
-import { useOrderListContext } from "../../../context/OrderListProvider";
-import { useOrderCalculation } from "../../../hooks/order/useOrderCalculation";
-import OrderLabelPridce from "./OrderLabelPrice";
+import { getShoppingCartData } from "../../../../api/cart";
+import { Flex } from "../../../../components/common";
+import CheckboxLabel from "../../../../components/common/CheckboxLabel";
+import { useAPIDataContext } from "../../../../context/APIDataProvider";
+import { useOrderListContext } from "../../../../context/OrderListProvider";
+import { useOrderCalculation } from "../../../../hooks/order/useOrderCalculation";
+import OrderLabelPridce from "../OrderLabelPrice";
 
 function DeliveryInfo() {
   const [isIsland, setIsIsland] = useState(false);
@@ -35,6 +35,7 @@ function DeliveryInfo() {
         totalCartPrice={totalCartPrice}
         shippingFee={shippingFee}
         totalPrice={totalPrice}
+        couponDiscount={0}
       />
     </Container>
   );
